tcp-sim cleanup
